Divides js into modules
